Rename options page folder to reflect its new content
