+++ package.json
@@ -7,8 +7,8 @@
     "buildServer": "tsc",
     "devServer": "nodemon startServer",
     "lint": "eslint src",
-    "buildUI": "vite build --config src/plugin/options_page/vite.config.ts",
-    "devUI": "vite --config src/plugin/options_page/vite.config.ts",
+    "buildUI": "vite build --config src/plugin/user_interface/vite.config.ts",
+    "devUI": "vite --config src/plugin/user_interface/vite.config.ts",
     "buildScriptsAndWrapPlugin": "webpack",
     "watchScriptsAndWrapPlugin": "webpack --watch",
     "buildPlugin": "npm run buildUI && npm run buildScriptsAndWrapPlugin"
