+++ src/plugin/user_interface/components/ManageSession.vue
