+++ src/plugin/user_interface/vite.config.ts
